requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
astroid==3.3.10
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
autopep8==1.5.5
backports.tarfile==1.2.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
crcmod==1.7
cryptography==45.0.5
dbus-python==1.3.2
decorator==5.2.1
dill==0.4.0
distlib==0.3.9
distro==1.9.0
Django==5.2.4
dnspython==2.7.0
docstring_parser==0.16
executing==2.2.0
filelock==3.18.0
flake8==3.8.4
Flask==3.1.1
Flask-PyMongo==3.0.1
flatbuffers==25.2.10
gast==0.6.0
gce-v2v @ https://storage.googleapis.com/velostrata-release/gce-v2v/gce-v2v.tar.gz#sha256=1533c75fd582af8102f3c7bb800342521b0964266f723e29ebd5afddc1089c4d
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.102.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-datastore==2.21.0
google-cloud-firestore==2.21.0
google-cloud-language==2.17.2
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-spanner==3.55.0
google-cloud-storage==2.19.0
google-cloud-translate==3.21.1
google-cloud-videointelligence==2.16.2
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-genai==1.25.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.17.2
jeepney==0.9.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
keyring==25.6.0
keyrings.google-artifactregistry-auth==1.1.2
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.6.1
mdurl==0.1.2
mercurial==6.7.2
ml_dtypes==0.5.1
more-itertools==10.7.0
namex==0.1.0
numpy==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
opentelemetry-api==1.34.1
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
parso==0.7.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==2.2.0
Pygments==2.19.2
PyGObject==3.48.2
PyJWT==2.7.0
pylint==3.3.7
pymongo==4.13.2
PyMuPDF==1.26.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytoolconfig==1.3.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rope==1.13.0
rpds-py==0.26.0
rsa==4.9.1
SecretStorage==3.3.3
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
sqlparse==0.5.3
stack-data==0.6.3
supervisor==4.2.5
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tink==1.11.0
toml==0.10.2
tomlkit==0.13.3
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.31.2
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yapf==0.43.0
zipp==3.23.0
